up to svn 61114

+++ Services/COPage/classes/class.ilPCFileList.php
@@ -11,7 +11,7 @@ require_once("./Services/COPage/classes/class.ilPageContent.php");
 * File List content object (see ILIAS DTD)
 *
 * @author Alex Killing <alex.killing@gmx.de>
-* @version $Id: class.ilPCFileList.php 60741 2015-09-17 08:53:32Z bheyser $
+* @version $Id: class.ilPCFileList.php 61113 2015-10-16 13:38:50Z bheyser $
 *
 * @ingroup ServicesCOPage
 */
@@ -261,22 +261,26 @@ class ilPCFileList extends ilPageContent
 	 */
 	static function afterPageUpdate($a_page, DOMDocument $a_domdoc, $a_xml, $a_creation)
 	{
-		// pc filelist
-		include_once("./Modules/File/classes/class.ilObjFile.php");
-		$file_ids = ilObjFile::_getFilesOfObject(
-			$a_page->getParentType().":pg", $a_page->getId(), 0, $a_page->getLanguage());
-		self::saveFileUsage($a_page, $a_domdoc);
-		foreach($file_ids as $file)	// check, whether file object can be deleted
+		if (!$a_page->getImportMode())
 		{
-			if (ilObject::_exists($file))
+			// pc filelist
+			include_once("./Modules/File/classes/class.ilObjFile.php");
+			$file_ids = ilObjFile::_getFilesOfObject(
+				$a_page->getParentType().":pg", $a_page->getId(), 0, $a_page->getLanguage());
+			self::saveFileUsage($a_page, $a_domdoc);
+
+			foreach($file_ids as $file)	// check, whether file object can be deleted
 			{
-				$file_obj = new ilObjFile($file, false);
-				$usages = $file_obj->getUsages();
-				if (count($usages) == 0)	// delete, if no usage exists
+				if (ilObject::_exists($file) && ilObject::_lookupType($file) == "file")
 				{
-					if ($file_obj->getMode() == "filelist")		// non-repository object
+					$file_obj = new ilObjFile($file, false);
+					$usages = $file_obj->getUsages();
+					if (count($usages) == 0)	// delete, if no usage exists
 					{
-						$file_obj->delete();
+						if ($file_obj->getMode() == "filelist")		// non-repository object
+						{
+							$file_obj->delete();
+						}
 					}
 				}
 			}
